requirements.txt
Flask
easyocr
transformers
torch
pandas
pillow
matplotlib
requests
